Compiler warning fixed & Bugfix
- Compiler warning fixed: unsigned cast warning
- Bugfix (RPARSER.BMSARCHIVE): Angelico[MAX] Longnote processing,
- Bugfix: BMS_HTML_EXPORT feature

+++ src/Chart.h
@@ -77,7 +77,7 @@ public:
   uint32_t GetScoreableNoteCount() const;
   double GetSongLastObjectTime() const;
   bool HasLongnote() const;
-  uint8_t GetPlayLaneCount() const;
+  unsigned GetPlayLaneCount() const;
   
   void Clear();
